Fix compiler warning
v_utils.c: In function ‘check_maxlen’:
v_utils.c:324:48: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ vgrabbj/vgrabbj.h
@@ -250,7 +250,7 @@ struct v_options {
   int var_type;
   int min_value;
   int max_value;
-  int max_length;
+  size_t max_length;
   const char *standard;
   struct v_options *next;
 };
